requirements.txt
Cython==0.20.1
beautifulsoup4==4.4.1
django-smuggler==0.6.0
gspread==0.2.5
oauth2client==1.4.11
boto==2.38.0
django-storages-redux==1.2.3
django-sslify>=0.2.0
honcho==0.6.6
dj-database-url==0.3.0
gunicorn==19.3.0
DateTime==4.0.1
Django==1.8.2
GitPython==0.3.6
Jinja2==2.7.3
MarkupSafe==0.18
Pillow==2.8.1
PyInstaller==2.1
PyYAML==3.11
SQLAlchemy==0.9.4
Scrapy==0.24.6
Unidecode==0.04.17
Werkzeug==0.9.6
XlsxWriter==0.5.5
amqp==1.4.6
anyjson==0.3.3
argcomplete==0.6.7
atom==0.3.7
backports.ssl-match-hostname==3.4.0.2
binstar==0.11.0
bitarray==0.8.1
cdecimal==2.3
celery==3.1.18
certifi==14.05.14
cffi==0.9.2
characteristic==14.3.0
colorama==0.3.3
configobj==5.0.5
configparser==3.5.0b2
cryptography==0.8.2
cssselect==0.9.1
cubes==0.10.2
distlib==0.2.0
django-admin-bootstrapped==2.5.1
django-appconf==1.0.1
django-bootstrap3==5.4.0
django-celery==3.1.16
django-compressor==1.5
django-constance-updated==1.0.1
django-crispy-forms==1.4.0
django-djconfig==0.4.0
django-geoposition==0.2.2
django-getenv==1.3.1
django-picklefield==0.3.1
django-pipeline==1.5.1
django-static-precompiler==0.9
django-tables2==1.0.4
docopt==0.6.1
enaml==0.9.1
enum34==1.0.4
future==0.14.1
futures==3.0.3
gevent==1.0.1
gevent-websocket==0.9.3
gitdb==0.6.4
greenlet==0.4.2
grin==1.2.1
ipython==2.1.0
itsdangerous==0.24
jdcal==1.0
kiwisolver==0.1.2
kombu==3.0.26
logging==0.4.9.6
lxml==3.4.0
meld3==1.0.2
mock==1.0.1
multipledispatch==0.4.3
networkx==1.8.1
nose==1.3.3
numpy==1.8.1
partial==1.0
patsy==0.2.1
pep8==1.5.6
ply==3.4
psutil==2.1.1
psycopg2==2.6.1
py==1.4.20
pyOpenSSL==0.15.1
pycurl==7.19.0
pyface==4.4.0
pyflakes==0.8.1
pygeocoder==1.2.5
pyparsing==2.0.1
pyreadline==2.0
pytest==2.5.2
python-dateutil==1.5
python-vlookup==1.0
pytkapp==0.3.70.3
pytz==2014.3
pyzmq==14.3.0
queuelib==1.2.2
redis==2.10.3
requests==2.7.0
rope==0.9.4
runipy==0.1.0
scrapple==0.2.4
selenium==2.45.0
service-identity==14.0.0
six==1.9.0
smmap==0.9.0
spyder==2.3.0
supervisor==3.1.3
tornado==3.2.1
tox==1.9.0
traits==4.4.0
traitsui==4.4.0
ujson==1.33
virtualenv==12.0.7
w3lib==1.11.0
wget==2.2
wsgiref==0.1.2
xlrd==0.9.3
xlwings==0.1.0
xlwt==0.7.5
zope.interface==4.1.2
